fix never clearing destination lists on each pipe block entity
made the destination lists include the destination node
